fix: make SveltexCache serializable for Vite
Vite was consistently throwing the following error when I tried to start
a dev server:

```
TypeError: Converting circular structure to JSON
```

The reason turned out to be that I had added a property `cacheDirGlob`
to the `SveltexCache`. Now, it seems that Vite serializes a lot of
stuff (I'm assuming at least partially this is for caching purposes),
and was trying to serialize the `SveltexCache` that `Sveltex` created.
However, `Glob` objects aren't always (ever?) serializable (or at least
that's the impression I've gotten), which was the cause of the error.
The solution: to add a custom `toJSON` method to the class that returns
the class without the `cacheDirGlob` property. The `toJSON` method is
then called by `JSON.strinfigy`, which will then stringify its output
(i.e., the `SveltexCache` instance without the `cacheDirGlob` prop)
instead of the full `SveltexCache` instance.

Why bother with a `cacheDirGlob` prop to begin with? Well, it gives us
significant performance gains when walking the cache directory; see
`glob`'s [readme](https://github.com/isaacs/node-glob#readme).