fix(deps): hard-code inquirer version (#9)
Issue #9 seems to be a consequence of
https://github.com/plopjs/plop/issues/445. As a temporary workaround, as
suggested in that issue's comments, I'm hard-coding the 9.3.5 version of
`inquirer` in `package.json`, which should fix the issue for now. If and
when Plop and/or Inquirer fix the root cause, this temporary fix here
should be reverted and a newer minimum version of `plop` should be
required in the `package.json` file.

Fixes #9.