fix: improve inquirer version lock (#9)
This commit tries to further hard-code the inquirer version in
package.json by falsely listing it as a direct dependency (instead of as
a peer dependency), in the hopes that this may fix the
`pnpm dlx create-sveltex` and `bunx create-sveltex` commands.

This commit also bumps the create-sveltex version number and updates the
changelog to prepare for publication.

Related issue: #9.