fix(#25): modernize element-detecting regexes
I'm not really sure why this fixes the issue, but it's what I could
trace it back to. Passing the regex returned by XRegExp to
`document.matchAll(...)` made the script tag not be detected if a curly
quote was present, whereas with the regex returned by the regex function
it worked. However, I was unable to reproduce this behavior in
isolation, so I also don't know if its a bug with XRegExp, or something
more confusing.